requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.2
Flask==2.0.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
PyJWT==2.2.0
pyngrok==5.1.0
python-aiml==0.9.3
python-dotenv==0.19.1
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
twilio==7.1.0
urllib3==1.26.7
Werkzeug==2.0.2
